requirements.txt
instagrapi==1.16.21
Pillow==9.2.0
requests==2.28.1
arabic-reshaper==2.1.3